refactor on artmesh transformations

+++ Assets/Scripts/Animation/ParamPoint.cs
@@ -5,13 +5,13 @@ public class ParamPoint
     private static ushort _paramPtCounter = 0;
 
     public readonly ushort ID;
-    public int ParamValue;
+    public float ParamValue;
 
     public Vector3 Position { get; set; }
     public Quaternion Rotation { get; set; }
     public Vector3 Scale { get; set; }
 
-    public ParamPoint(int paramValue, bool autoRegister = true)
+    public ParamPoint(float paramValue, bool autoRegister = true)
     {
         ID = _paramPtCounter++;
         this.ParamValue = paramValue;
